Revert "Added files: Landing, logoImg; Changed routing to include Landing; Created first verion of Landing page"
This reverts commit 56d1ec0d6a0cae87e22fd0fccab203367fab9df4.
