+++ public/index.html
@@ -23,7 +23,7 @@
       Learn how to configure a non-root public URL by running `npm run build` or
       have a look at the documentation https://create-react-app.dev/docs/using-the-public-folder/.
     -->
-    <title>SoPra FS24 - Group 07</title>
+    <title>SoPra FS24 - Group XX</title>
   </head>
 
   <body>
