+++ src/styles/_theme.scss
@@ -8,10 +8,10 @@
    feel free to play around with these and see what other look you can come up with! :D
 */
 
-$accent: #cdecf9;
-$accentDark: #cdecf9;
+$accent: #97abff;
+$accentDark: #123597;
 $textColor: white;
-$background: #02152a;
+$background: #262632;
 
 /* some constants */
