+++ src/styles/ui/BaseContainer.scss
@@ -6,5 +6,5 @@
   margin-right: auto;
   padding-right: 15px;
   max-width: $DESKTOP_WIDTH;
-  color: $accent;
+  color: $textColor;
 }
