+++ src/styles/ui/Button.scss
@@ -16,7 +16,7 @@
   text-transform: uppercase;
   font-size: 13px;
   text-align: center;
-  height: 10px;
+  height: 35px;
   border: none;
   border-radius: $borderRadius;
   cursor: pointer;
