--- src/styles/ui/Logo.scss
@@ -1,5 +1,0 @@
-@import "../theme";
-
-.logo-img {
-  margin: 10px; /* Adjust the margin value as needed */
-}
