+++ src/styles/views/Login.scss
@@ -2,17 +2,11 @@
 
 .login {
   &.container {
-    margin-top: 1em;
+    margin-top: 2em;
+    min-height: 300px;
+    /* we include the center mixin defined in _theme.scss and save three duplicate lines :) */
+    @include center;
     flex-direction: column;
-    display: flex; /* Add this to enable flexbox */
-    justify-content: center; /* Center vertically */
-    align-items: center; /* Center horizontally */
-  }
-
-  &.text {
-    color: $background;
-    text-align: center;
-    margin-top: 35em;
   }
   &.form {
     display: flex;
@@ -20,7 +14,7 @@
     justify-content: center;
     width: 60%;
     max-width: 30em;
-    height: 0em;
+    height: 20em;
     font-size: 16px;
     font-weight: 300;
     padding-left: 37px;
@@ -43,19 +37,19 @@
     margin-left: -4px;
     border: none;
     border-radius: 0.75em;
-    margin-bottom: 1em;
+    margin-bottom: 20px;
     background: transparentize(white, 0.85);
     color: $textColor;
   }
   &.label {
     color: $textColor;
-    margin-bottom: 1em;
+    margin-bottom: 10px;
     text-transform: lowercase;
     font-weight: 300;
   }
   &.button-container {
     display: flex;
     justify-content: center;
-    margin-top: 1em;
+    margin-top: 2em;
   }
 }
